+++ src/benchmark/benchmark_sigmoid.c
@@ -13,9 +13,9 @@ int main() {
         printf("Running benchmark_sigmoid\n");
 
         // x is input; result is output of the optimized functions
-        float* x = allocate(size, sizeof(float));
-        float* result_ref = allocate(size, sizeof(float));
-        float* result = allocate(size, sizeof(float));
+        double* x = allocate(size, sizeof(double));
+        double* result_ref = allocate(size, sizeof(double));
+        double* result = allocate(size, sizeof(double));
 
         srandom(2);
         x[0] = 0.0; // sigmoid(0.0) is 0.5
