+++ src/lmq/lmq.h
@@ -22,12 +22,12 @@ void* allocate(const size_t n, const size_t element_size);
 /*
  * Prints a matrix which is in the form [rows x columns] and row major inside arr.
  */
-void print_matrix(const float* arr, const size_t rows, const size_t cols);
+void print_matrix(const double* arr, const size_t rows, const size_t cols);
 
 /*
  * Calculates an approximation of the square root of a.
  * Needed as the fsqrt instruction is not implemented on the snitch. 
  */
-float sqrt_approx(float a);
+double sqrt_approx(double a);
 
 #endif
